move fake catalog next to real catalog
